Extract `format_byte_size` from `fetch_tracer`
Bug: b/535754069
