+++ base/cvd/cuttlefish/host/commands/cvd/cli/format_byte_size.cc
@@ -1,0 +1,46 @@
+//
+// Copyright (C) 2024 The Android Open Source Project
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "cuttlefish/host/commands/cvd/cli/format_byte_size.h"
+
+#include <stdint.h>
+
+#include <string>
+
+#include "fmt/format.h"
+
+namespace cuttlefish {
+
+std::string FormatByteSize(uint64_t size) {
+  if (size < 10240) {
+    return fmt::format("{} B", size);
+  }
+  size /= 1024;
+  if (size < 10240) {
+    return fmt::format("{} KiB", size);
+  }
+  size /= 1024;
+  if (size < 10240) {
+    return fmt::format("{} MiB", size);
+  }
+  size /= 1024;
+  if (size < 10240) {
+    return fmt::format("{} GiB", size);
+  }
+  size /= 1024;
+  return fmt::format("{} TiB", size);
+}
+
+}  // namespace cuttlefish
